can reference data from the ws server to the server

+++ db/migrations/20181121202312_locations.js
@@ -2,8 +2,8 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('locations', (table) => {
     table.increments('id').primary();
     table.integer('user_id').unsigned().notNull();
-    table.integer('lat');
-    table.integer('long');
+    table.float('lat');
+    table.float('long');
     table.dateTime('last_check_in')
 
     table.foreign('user_id').references('id').inTable('users');
